+++ src/lib/components/modal/code/CodeDetailsTemplate.tsx
@@ -5,7 +5,7 @@ import { MdAddCircleOutline, MdCheckCircle } from "react-icons/md";
 import { ActionModal } from "..";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { TextInput } from "lib/components/forms";
-import { MAX_CODE_DESCRIPTION_LENGTH } from "lib/data";
+import { MAX_CODE_NAME_LENGTH } from "lib/data";
 import { useCodeStore, useGetAddressType } from "lib/hooks";
 import type { CodeLocalInfo } from "lib/stores/code";
 
@@ -30,16 +30,14 @@ export const CodeDetailsTemplateModal = ({
   const toast = useToast();
   const getAddressType = useGetAddressType();
 
-  const [description, setDescription] = useState(
-    codeLocalInfo.description ?? ""
-  );
+  const [name, setName] = useState(codeLocalInfo.name ?? "");
 
   const uploaderType = getAddressType(codeLocalInfo.uploader);
 
   const handleAction = useCallback(() => {
     if (isNewCode) saveNewCode(codeLocalInfo.id);
 
-    updateCodeInfo(codeLocalInfo.id, codeLocalInfo.uploader, description);
+    updateCodeInfo(codeLocalInfo.id, codeLocalInfo.uploader, name);
 
     // TODO: abstract toast to template later
     toast({
@@ -61,7 +59,7 @@ export const CodeDetailsTemplateModal = ({
   }, [
     codeLocalInfo.id,
     codeLocalInfo.uploader,
-    description,
+    name,
     isNewCode,
     saveNewCode,
     title,
@@ -69,10 +67,10 @@ export const CodeDetailsTemplateModal = ({
     updateCodeInfo,
   ]);
 
-  // fix prefilling blank space problem (e.g. description saved as "     ")
+  // fix prefilling blank space problem (e.g. name saved as "     ")
   useEffect(() => {
-    setDescription(codeLocalInfo.description ?? "");
-  }, [codeLocalInfo.description]);
+    setName(codeLocalInfo.name ?? "");
+  }, [codeLocalInfo.name]);
 
   return (
     <ActionModal
@@ -113,13 +111,13 @@ export const CodeDetailsTemplateModal = ({
     >
       <TextInput
         variant="floating"
-        value={description}
-        setInputState={setDescription}
+        value={name}
+        setInputState={setName}
         size="lg"
-        helperText="Fill in code description to define its use as a reminder"
-        label="Code Description"
+        helperText="Fill in code name to define its use as a reminder"
+        label="Code Name"
         labelBgColor="pebble.900"
-        maxLength={MAX_CODE_DESCRIPTION_LENGTH}
+        maxLength={MAX_CODE_NAME_LENGTH}
       />
     </ActionModal>
   );