+++ src/lib/model/code.ts
@@ -6,17 +6,36 @@ import {
   useContractListByCodeId,
   useContractListCountByCodeId,
 } from "lib/services/codeService";
+import {
+  usePublicProjectByCodeId,
+  usePublicProjectBySlug,
+} from "lib/services/publicProjectService";
 import type { ContractLocalInfo } from "lib/stores/contract";
-import type { CodeData, ContractInstances, Option } from "lib/types";
+import type {
+  CodeData,
+  ContractInstances,
+  PublicDetail,
+  Option,
+  PublicCodeData,
+} from "lib/types";
 
 interface CodeDataState {
   isLoading: boolean;
   codeData: CodeData;
+  publicProject: {
+    publicCodeData: Option<PublicCodeData>;
+    publicDetail: Option<PublicDetail>;
+  };
 }
 
 export const useCodeData = (codeId: number): Option<CodeDataState> => {
   const { currentChainRecord } = useWallet();
   const { data: codeInfo, isLoading } = useCodeInfoByCodeId(codeId);
+  const { data: publicCodeInfo } = usePublicProjectByCodeId(codeId);
+  const { data: publicInfoBySlug } = usePublicProjectBySlug(
+    publicCodeInfo?.slug
+  );
+
   if (!currentChainRecord || (!codeInfo && !isLoading)) return undefined;
 
   return {
@@ -25,6 +44,10 @@ export const useCodeData = (codeId: number): Option<CodeDataState> => {
       chainId: currentChainRecord.chain.chain_id,
       ...codeInfo,
     } as CodeData,
+    publicProject: {
+      publicCodeData: publicCodeInfo,
+      publicDetail: publicInfoBySlug?.details,
+    },
   };
 };